migrate existing catalyze tests
