+++ internal/tests/reactable/catalyze_test.go
@@ -10,6 +10,43 @@ import (
 	"github.com/genshinsim/gcsim/pkg/reactable"
 )
 
+func TestQuicken(t *testing.T) {
+	c, _ := makeCore(1)
+	err := c.Init()
+	if err != nil {
+		t.Errorf("error initializing core: %v", err)
+		t.FailNow()
+	}
+	c.QueueAttackEvent(&combat.AttackEvent{
+		Info: combat.AttackInfo{
+			Element:    attributes.Dendro,
+			Durability: 25,
+		},
+		Pattern: combat.NewCircleHit(combat.NewCircle(0, 0, 1), 100, false, combat.TargettableEnemy),
+	}, 0)
+	advanceCoreFrame(c)
+	c.QueueAttackEvent(&combat.AttackEvent{
+		Info: combat.AttackInfo{
+			Element:    attributes.Electro,
+			Durability: 25,
+		},
+		Pattern: combat.NewCircleHit(combat.NewCircle(0, 0, 1), 100, false, combat.TargettableEnemy),
+	}, 0)
+	advanceCoreFrame(c)
+
+	trg := c.Combat.Enemies()[0].(*enemy.Enemy)
+	if math.Abs(float64(trg.Durability[reactable.ModifierQuicken])-19.966666) > 0.000001 {
+		t.Errorf(
+			"expected quicken=%v, got quicken=%v",
+			19.666666,
+			trg.Durability[reactable.ModifierQuicken],
+		)
+	}
+	if trg.AuraContains(attributes.Dendro, attributes.Electro) {
+		t.Error("expecting target to not contain any remaining dendro or electro aura")
+	}
+}
+
 func TestElectroDoesNotReduceQuicken(t *testing.T) {
 	c, _ := makeCore(1)
 	err := c.Init()
